Add files via upload
This is the sql script to create our database. The .mwb is to use with mysql workbench and that provides a visual representation of the DB. In order to load the DB you have to have the mysql suite installed, this includes workbench and server. Its free to download: https://dev.mysql.com/downloads/installer/